+++ src/solverpy/builder/enigma.py
@@ -7,7 +7,7 @@ from .builder import Builder, NAME
 from .autotuner import AutoTuner
 from ..benchmark.path import sids, bids
 from .plugins import enigma
-from ..benchmark.setups.setup import Setup
+from ..setups.setup import Setup
 
 logger = logging.getLogger(__name__)